[Core][Order] Huge amount of fixes to make Sylius work
