+++ src/Sylius/Bundle/CoreBundle/Form/Type/CartItemType.php
@@ -11,9 +11,11 @@
 
 namespace Sylius\Bundle\CoreBundle\Form\Type;
 
-use Sylius\Bundle\CartBundle\Form\Type\CartItemType as BaseCartItemType;
+use Sylius\Bundle\OrderBundle\Form\Type\OrderItemType as BaseOrderItemType;
+use Sylius\Bundle\ResourceBundle\Form\Type\AbstractResourceType;
 use Sylius\Component\Core\Model\Product;
 use Sylius\Component\Core\Model\ProductInterface;
+use Symfony\Component\Form\DataMapperInterface;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
@@ -25,14 +27,37 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
  *
  * @author Paweł Jędrzejewski <pawel@sylius.org>
  */
-class CartItemType extends BaseCartItemType
+class CartItemType extends AbstractResourceType
 {
+    /**
+     * @var DataMapperInterface
+     */
+    protected $orderItemQuantityDataMapper;
+
+    /**
+     * @param string $dataClass
+     * @param array $validationGroups
+     * @param DataMapperInterface $orderItemQuantityDataMapper
+     */
+    public function __construct($dataClass, array $validationGroups = [], DataMapperInterface $orderItemQuantityDataMapper)
+    {
+        parent::__construct($dataClass, $validationGroups);
+
+        $this->orderItemQuantityDataMapper = $orderItemQuantityDataMapper;
+    }
+
     /**
      * {@inheritdoc}
      */
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
-        parent::buildForm($builder, $options);
+        $builder
+            ->add('quantity', 'integer', [
+                'attr' => ['min' => 1],
+                'label' => 'sylius.ui.quantity',
+            ])
+            ->setDataMapper($this->orderItemQuantityDataMapper)
+        ;
 
         if (isset($options['product']) && $options['product']->hasVariants() && !$options['product']->isSimple()) {
             $type = Product::VARIANT_SELECTION_CHOICE === $options['product']->getVariantSelectionMethod() ? 'sylius_product_variant_choice' : 'sylius_product_variant_match';
@@ -60,5 +85,13 @@ class CartItemType extends BaseCartItemType
             ])
             ->setAllowedTypes('product', ProductInterface::class)
         ;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getName()
+    {
+        return 'sylius_cart_item';
     }
 }
