+++ src/Sylius/Bundle/CoreBundle/Resources/translations/flashes.en.yml
@@ -1,5 +1,11 @@
 sylius:
     cart:
+        cart_clear_completed: The cart has been successfully cleared.
+        cart_save_completed: The cart has been successfully updated.
+        item_add_completed: Item has been added to cart.
+        item_add_error: Error occurred while adding item to cart.
+        item_remove_completed: Item has been removed from cart.
+        item_remove_error: Error occurred while removing item from cart.
         restricted_zone_removal: Product %product% removed from cart since it is not available in your country.
     checkout:
         restricted_zone: This product is not available in your country.
