--- src/Sylius/Bundle/OrderBundle/Controller/CartController.php
@@ -1,147 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Paweł Jędrzejewski
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Sylius\Bundle\OrderBundle\Controller;
-
-use FOS\RestBundle\View\View;
-use Sylius\Component\Order\SyliusCartEvents;
-use Sylius\Component\Resource\ResourceActions;
-use Symfony\Component\EventDispatcher\GenericEvent;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\Exception\HttpException;
-
-/**
- * @author Paweł Jędrzejewski <pawel@sylius.org>
- * @author Grzegorz Sadowski <grzegorz.sadowski@lakion.com>
- */
-class CartController extends Controller
-{
-    /**
-     * @param Request $request
-     *
-     * @return Response
-     */
-    public function summaryAction(Request $request)
-    {
-        $configuration = $this->requestConfigurationFactory->create($this->metadata, $request);
-
-        $cart = $this->getCurrentCart();
-        $form = $this->resourceFormFactory->create($configuration, $cart);
-
-        $view = View::create()
-            ->setTemplate($configuration->getTemplate('summary.html'))
-            ->setData([
-                'cart' => $cart,
-                'form' => $form->createView(),
-            ])
-        ;
-
-        return $this->viewHandler->handle($configuration, $view);
-    }
-
-    /**
-     * @param Request $request
-     *
-     * @return Response
-     */
-    public function saveAction(Request $request)
-    {
-        $configuration = $this->requestConfigurationFactory->create($this->metadata, $request);
-
-        $this->isGrantedOr403($configuration, ResourceActions::UPDATE);
-        $resource = $this->getCurrentCart();
-
-        $form = $this->resourceFormFactory->create($configuration, $resource);
-
-        if (in_array($request->getMethod(), ['POST', 'PUT', 'PATCH']) && $form->submit($request, !$request->isMethod('PATCH'))->isValid()) {
-            $resource = $form->getData();
-
-            $event = $this->eventDispatcher->dispatchPreEvent(ResourceActions::UPDATE, $configuration, $resource);
-
-            if ($event->isStopped() && !$configuration->isHtmlRequest()) {
-                throw new HttpException($event->getErrorCode(), $event->getMessage());
-            }
-            if ($event->isStopped()) {
-                $this->flashHelper->addFlashFromEvent($configuration, $event);
-
-                return $this->redirectHandler->redirectToResource($configuration, $resource);
-            }
-
-            if ($configuration->hasStateMachine()) {
-                $this->stateMachine->apply($configuration, $resource);
-            }
-
-            $this->eventDispatcher->dispatchPostEvent(ResourceActions::UPDATE, $configuration, $resource);
-
-            if (!$configuration->isHtmlRequest()) {
-                return $this->viewHandler->handle($configuration, View::create(null, Response::HTTP_NO_CONTENT));
-            }
-
-            $this->getEventDispatcher()->dispatch(SyliusCartEvents::CART_CHANGE, new GenericEvent($resource));
-            $this->manager->flush();
-
-            $this->flashHelper->addSuccessFlash($configuration, ResourceActions::UPDATE, $resource);
-
-            return $this->redirectHandler->redirectToResource($configuration, $resource);
-        }
-
-        if (!$configuration->isHtmlRequest()) {
-            return $this->viewHandler->handle($configuration, View::create($form, Response::HTTP_BAD_REQUEST));
-        }
-
-        $view = View::create()
-            ->setData([
-                'configuration' => $configuration,
-                $this->metadata->getName() => $resource,
-                'form' => $form->createView(),
-            ])
-            ->setTemplate($configuration->getTemplate(ResourceActions::UPDATE . '.html'))
-        ;
-
-        return $this->viewHandler->handle($configuration, $view);
-    }
-
-    /**
-     * @param Request $request
-     *
-     * @return Response
-     */
-    public function clearAction(Request $request)
-    {
-        $configuration = $this->requestConfigurationFactory->create($this->metadata, $request);
-
-        $this->isGrantedOr403($configuration, ResourceActions::DELETE);
-        $resource = $this->getCurrentCart();
-
-        $event = $this->eventDispatcher->dispatchPreEvent(ResourceActions::DELETE, $configuration, $resource);
-
-        if ($event->isStopped() && !$configuration->isHtmlRequest()) {
-            throw new HttpException($event->getErrorCode(), $event->getMessage());
-        }
-        if ($event->isStopped()) {
-            $this->flashHelper->addFlashFromEvent($configuration, $event);
-
-            return $this->redirectHandler->redirectToIndex($configuration, $resource);
-        }
-
-        $this->repository->remove($resource);
-        $this->eventDispatcher->dispatchPostEvent(ResourceActions::DELETE, $configuration, $resource);
-
-        if (!$configuration->isHtmlRequest()) {
-            return $this->viewHandler->handle($configuration, View::create(null, Response::HTTP_NO_CONTENT));
-        }
-
-        $this->flashHelper->addSuccessFlash($configuration, ResourceActions::DELETE, $resource);
-
-        return $this->redirectHandler->redirectToIndex($configuration, $resource);
-    }
-}
