--- src/Sylius/Bundle/OrderBundle/Controller/Controller.php
@@ -1,71 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Paweł Jędrzejewski
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Sylius\Bundle\OrderBundle\Controller;
-
-use Sylius\Bundle\ResourceBundle\Controller\RequestConfiguration;
-use Sylius\Bundle\ResourceBundle\Controller\ResourceController;
-use Sylius\Component\Order\Context\CartContextInterface;
-use Sylius\Component\Order\Model\OrderInterface;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
-use Symfony\Component\HttpFoundation\RedirectResponse;
-
-/**
- * @author Paweł Jędrzejewski <pawel@sylius.org>
- */
-abstract class Controller extends ResourceController
-{
-    /**
-     * @param RequestConfiguration $configuration
-     *
-     * @return RedirectResponse
-     */
-    protected function redirectToCartSummary(RequestConfiguration $configuration)
-    {
-        if (null === $configuration->getParameters()->get('redirect')) {
-            return $this->redirectHandler->redirectToRoute($configuration, $this->getCartSummaryRoute());
-        }
-
-        return $this->redirectHandler->redirectToRoute($configuration, $configuration->getParameters()->get('redirect'));
-    }
-
-    /**
-     * @return string
-     */
-    protected function getCartSummaryRoute()
-    {
-        return 'sylius_cart_summary';
-    }
-
-    /**
-     * @return OrderInterface
-     */
-    protected function getCurrentCart()
-    {
-        return $this->getContext()->getCart();
-    }
-
-    /**
-     * @return CartContextInterface
-     */
-    protected function getContext()
-    {
-        return $this->container->get('sylius.context.cart');
-    }
-
-    /**
-     * @return EventDispatcherInterface
-     */
-    protected function getEventDispatcher()
-    {
-        return $this->container->get('event_dispatcher');
-    }
-}
