+++ src/Sylius/Bundle/OrderBundle/Resources/config/routing.yml
@@ -3,20 +3,20 @@
 
 sylius_cart_summary:
     path: /
-    defaults: { _controller: sylius.controller.cart:summaryAction }
+    defaults: { _controller: sylius.controller.order:summaryAction }
 
 sylius_cart_save:
     path: /save
-    defaults: { _controller: sylius.controller.cart:saveAction }
+    defaults: { _controller: sylius.controller.order:saveAction }
 
 sylius_cart_clear:
     path: /clear
-    defaults: { _controller: sylius.controller.cart:clearAction }
+    defaults: { _controller: sylius.controller.order:clearAction }
 
 sylius_cart_item_add:
     path: /add
-    defaults: { _controller: sylius.controller.cart_item:addAction }
+    defaults: { _controller: sylius.controller.order_item:addAction }
 
 sylius_cart_item_remove:
     path: /{id}/remove
-    defaults: { _controller: sylius.controller.cart_item:removeAction }
+    defaults: { _controller: sylius.controller.order_item:removeAction }
