+++ src/Sylius/Bundle/OrderBundle/Resources/config/services.xml
@@ -75,13 +75,13 @@
             <argument type="service" id="sylius.order_item_quantity_modifier" />
         </service>
         <service id="sylius.context.cart_new" class="Sylius\Component\Order\Context\CartContext">
-            <argument type="service" id="sylius.factory.cart" />
+            <argument type="service" id="sylius.factory.order" />
             <tag name="sylius.cart_context" priority="-999" />
         </service>
         <service id="sylius.context.cart_session_based" class="Sylius\Bundle\OrderBundle\Context\SessionBasedCartContext">
             <argument type="service" id="session" />
             <argument>_sylius.cart</argument>
-            <argument type="service" id="sylius.repository.cart" />
+            <argument type="service" id="sylius.repository.order" />
             <tag name="sylius.cart_context" priority="-777" />
         </service>
         <service id="sylius.context.cart" class="Sylius\Component\Order\Context\CompositeCartContext" />
