+++ src/Sylius/Bundle/ShopBundle/Resources/config/routing/cart.yml
@@ -5,15 +5,15 @@ sylius_shop_cart_summary:
     path: /
     methods: [GET]
     defaults:
-        _controller: sylius.controller.cart:summaryAction
+        _controller: sylius.controller.order:summaryAction
         _sylius:
             template: SyliusShopBundle:Cart:summary.html.twig
 
 sylius_shop_cart_save:
     path: /
-    methods: [POST]
+    methods: [POST, PUT, PATCH]
     defaults:
-        _controller: sylius.controller.cart:saveAction
+        _controller: sylius.controller.order:saveAction
         _sylius:
             template: SyliusShopBundle:Cart:summary.html.twig
             redirect: sylius_shop_cart_summary
@@ -23,14 +23,14 @@ sylius_shop_cart_clear:
     path: /
     methods: [DELETE]
     defaults:
-        _controller: sylius.controller.cart:clearAction
+        _controller: sylius.controller.order:clearAction
         _sylius:
             redirect: sylius_shop_cart_summary
 
 sylius_shop_cart_item_remove:
     path: /{id}/remove
     defaults:
-        _controller: sylius.controller.cart_item:removeAction
+        _controller: sylius.controller.order_item:removeAction
         _sylius:
             redirect: sylius_shop_cart_summary
             flash: sylius.cart.item_remove_completed
@@ -39,7 +39,7 @@ sylius_shop_cart_item_add:
     path: /add
     methods: [GET, POST]
     defaults:
-        _controller: sylius.controller.cart_item:addAction
+        _controller: sylius.controller.order_item:addAction
         _sylius:
             factory:
                 method: createForProduct
