+++ src/Sylius/Bundle/ShopBundle/Resources/views/Cart/Summary/_items.html.twig
@@ -1,6 +1,7 @@
 <div class="ui segment">
     {{ form_start(form, {'action': path('sylius_shop_cart_save'), 'attr': {'class': 'ui loadable form', 'novalidate': 'novalidate'}}) }}
     {{ form_errors(form) }}
+    <input type="hidden" name="_method" value="PATCH"/>
     <table id="sylius-cart-items" class="ui very basic celled table">
         <thead>
         <tr>
