+++ src/Sylius/Bundle/ShopBundle/Resources/views/Product/Show/_addToCart.html.twig
@@ -1,4 +1,4 @@
-{% set product = cart_item.variant.product %}
+{% set product = order_item.variant.product %}
 
 {% form_theme form 'SyliusUiBundle:Form:theme.html.twig' %}
